requirements.txt
openai~=0.27.2
python-dotenv~=1.0.0
discord.py